Refactor remaining plugin conversion tests out of MonitorConversionServiceTest
